+++ queosk/src/main/java/com/bttf/queosk/dto/enumerate/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.dto;
+package com.bttf.queosk.dto.enumerate;
 
 public enum OrderStatus {
     IN_PROGRESS, DONE, CANCELED
