+++ queosk/src/main/java/com/bttf/queosk/dto/enumerate/RestaurantCategory.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.dto;
+package com.bttf.queosk.dto.enumerate;
 
 public enum RestaurantCategory {
     DONKASU_SUSHI_JAPANESE,
