+++ queosk/src/main/java/com/bttf/queosk/entity/Order.java
@@ -1,7 +1,7 @@
 package com.bttf.queosk.entity;
 
 import com.bttf.queosk.config.BaseTimeEntity;
-import com.bttf.queosk.dto.OrderStatus;
+import com.bttf.queosk.dto.enumerate.OrderStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
